Service and Service status update

+++ database/migrations/2022_09_06_101011_create_user_services_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('user_services', function (Blueprint $table) {
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->unsignedInteger('service_id');
+            $table->unsignedSmallInteger('service_id');
             $table->foreign('service_id')->references('id')->on('services')->onDelete('cascade');
             $table->tinyInteger('quantity');
             $table->float('price');
